Moving TestUtils to separate file; Adding ValidateMachineInfoString() test which fails
