+++ Tests/ParallelBuildsMonitorTests.csproj
@@ -63,6 +63,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SaveCsvTests.cs" />
     <Compile Include="SavePngTests.cs" />
+    <Compile Include="TestUtils.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Key.snk" />
