+++ Tests/SavePngTests.cs
@@ -1,17 +1,11 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using ParallelBuildsMonitor;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IO;
 using System.Windows;
-
-using CM = ParallelBuildsMonitor.Tests.ComparisonMethods;
-using System.Windows.Controls;
 using System.Drawing;
 using System.Security.Cryptography;
+using ParallelBuildsMonitorTests;
 
 namespace ParallelBuildsMonitor.Tests
 {
